NVMe: successfully submitting commands
After some head scratching I have finally managed to submit a command to
the admin queue and get a seemingly proper result back (beats the
typical left overs of a pointer to the next freelist entry in the PMM).
I need to implement the second half of submitting the command, which is
acknowledging that the controller has completed it. The structures have
also been expanded in order to save myself eons of bitshifts and bitwise
ANDs (I could just use macros for this purpose, but I find this nicer).