NVMe: restructured
I have restructured how NVMe works. There is now a header file just for
definitions. These header files should not be used by any other parts of
the kernel that are unrelated to the given driver. I have improved the
initialization steps, and have separated the admin queue pair from the
rest so that it can be quickly found.