NVMe: masked reads, writes, creation of IO qpairs
I have replaced the packed structures with bit fields for macros that do
masked reads and masked writes. This is because the NVMe controller is
not too big a fan of writing on non 4 byte or 8 byte aligned
addresses (atleast for configuration stuff). With this, I have
successfully created the first IO qpair. The code does require a bit of
refactoring.