Created new tool to generate dri_defs
A python tool, which may in the future be translated to C, has been
created to auotmatically make dri_defs.h and dri_defs.c. This allows for
a look up table to be generated along with the definitions allowing for
O(1) lookup of driver definitions. The source code of the tool is very
rough, but it works. The output of it can also be a little bit cleaner.