Ext2 driver, hints
Work has began on the ext2 driver, the superblock and block group
descriptor table have been read in and stored in a driver state. Drivers
have also been made compatible with the new hint parameter. This should
allow super block drivers for filesystems on a drive to lookup nodes
faster.