Organized directories, added directory drivers
With the optimization of the size of driver definitions and resources,
I have added directory drivers. These are optional; however, they serve
to allow super block driver to not have to account for filepaths, rather
it can just account for filenames. The directory drivers will also
account for only filenames. This should improve speeds as now filepaths
no longer have to be traversed, and nodes already existing in memory
leveraged. The python tool to generate driver definitions should also
have a mechanism to optimize for space such that if a driver contains
all empty functions, then it resorts to just a default definition.