Re-organization, NVMe NS fix, EXT2 basics
The file structure has once more been re-organized. This is more logical
as folders now separate the groups of super, directory, and file drivers
in sysfs. This should make the use of header files and connecting like
drivers a bit easier. This was already done on the sysdev side. A
crucial fix has been implemented in the NVMe namespace driver. This
essentially corrects the jank that is introduced by NVMe being able to
only read on lba_size aligned offsets. The absence of this correction
caused errors when trying to implement the EXT2 driver. Some of the
basics of the EXT2 filesystem driver have been implemented (i.e. finding
inodes in the filesystem, reading them into memory, statting them, and
locating inodes to a depth of 1).