EXT2: revise headers, generalize block traversal
I have revised the header structure of the EXT2 driver, but more
importantly implemented a callback based, iterative block traverser such
that to implement reading and writing only a few functions have to be
implemented and given to a general one. This function has not yet been
tested for very large files (i.e. ones that need doubly or triply
indirect block pointers).