EXT2: implement directories
Basically a combination of super.c and file.c. Need to implement writing
to the VFS, this will be done by parsing the attributes given by the
superblock, as well as defining the control functions to allocate and
free a block in the VFS. Along with the ones to associate and
disassociate inodes from directories.