EXT2: implement checks, work on control function
I have implemented a checking function that will look over the
filesystem and, currently, properly determine if write enable should be
set. I have also been working on the control function, though it has
been a bit sluggish as there isn't a proper structure defined for it.
However, reading does work and I will be focusing my attention over to
getting a program running. After that I will come back and fix this up
such that it works for writing.